Move all coords-related utils to a separate file
